requirements.txt
coloredlogs==15.0.1
discord.py==2.2.2
SQLAlchemy==2.0.4
